inceput probelma 1

+++ src/tourists/Hotel.java
@@ -4,7 +4,7 @@ public class Hotel {
     private int availableRooms;
     private double price;
     private String name;
-    private int rating;
+    private double rating;
 
     public int getAvailableRooms() {
         return availableRooms;
@@ -30,11 +30,12 @@ public class Hotel {
         this.name = name;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(double rating) {
         this.rating = rating;
     }
+
 }
